+++ sdk/cloudmachine/Azure.Projects.Provisioning/src/AIProject/AIFoundryConnections.cs
