+++ sdk/cloudmachine/Azure.Projects.Provisioning/src/FeaturesBuiltIn/ServiceBusNamespaceFeature.cs
@@ -1,12 +1,11 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System.ClientModel.Primitives;
 using System.Collections.Generic;
 using Azure.Projects.Core;
-using Azure.Core;
 using Azure.Provisioning.Primitives;
 using Azure.Provisioning.ServiceBus;
-using System.ClientModel.Primitives;
 
 namespace Azure.Projects.ServiceBus;
