+++ sdk/cloudmachine/Azure.Projects.Provisioning/src/FeaturesBuiltIn/StorageAccountFeature.cs
@@ -1,14 +1,13 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System.ClientModel.Primitives;
 using System.Collections.Generic;
 using Azure.Projects.Core;
-using Azure.Core;
 using Azure.Provisioning.Expressions;
 using Azure.Provisioning.Primitives;
 using Azure.Provisioning.Resources;
 using Azure.Provisioning.Storage;
-using System.ClientModel.Primitives;
 
 namespace Azure.Projects.Storage;
