+++ sdk/cloudmachine/Azure.Projects.Provisioning/src/FeaturesExtensions/AIProjectFeature_internal.cs
@@ -1,10 +1,10 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Collections.Generic;
 using Azure.Provisioning.Expressions;
 using Azure.Provisioning.Primitives;
 using Azure.Provisioning.Roles;
-using System.Collections.Generic;
 
 namespace Azure.Provisioning.AIFoundry;
