--- sdk/cloudmachine/Azure.Projects.Provisioning/src/FeaturesExtensions/AgentsFeature.cs
@@ -1,17 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using Azure.Projects.Core;
-using Azure.Provisioning.Primitives;
-using System;
-using Azure.Projects;
-
-namespace Azure.AI.Agents;
-
-public class AgentsFeature : AzureProjectFeature
-{
-    protected override ProvisionableResource EmitResources(ProjectInfrastructure infrastructure)
-    {
-        throw new NotImplementedException();
-    }
-}
