+++ sdk/cloudmachine/Azure.Projects.Provisioning/src/FeaturesExtensions/AppServiceFeature.cs
@@ -1,11 +1,11 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using Azure.Provisioning.Expressions;
+using Azure.Projects.Core;
 using Azure.Provisioning.AppService;
+using Azure.Provisioning.Expressions;
 using Azure.Provisioning.Primitives;
 using Azure.Provisioning.Resources;
-using Azure.Projects.Core;
 
 namespace Azure.Projects.AppService;
