+++ sdk/cloudmachine/Azure.Projects.Provisioning/src/ProjectInfrastructure/ProjectInfrastructure.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System.ClientModel.Primitives;
 using System.Collections.Generic;
 using System.ComponentModel;
 using Azure.Projects.Core;
@@ -11,7 +12,6 @@ using Azure.Provisioning;
 using Azure.Provisioning.Expressions;
 using Azure.Provisioning.Primitives;
 using Azure.Provisioning.Roles;
-using System.ClientModel.Primitives;
 
 namespace Azure.Projects;
