+++ sdk/cloudmachine/Azure.Projects.Provisioning/src/ProjectInfrastructure/ProjectInfrastructure_plumbing.cs
@@ -1,21 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
-using System.Collections.Generic;
 using System.ComponentModel;
-using Azure.Projects.Core;
-using Azure.Projects.EventGrid;
-using Azure.Projects.ServiceBus;
-using Azure.Projects.Storage;
-using Azure.Core;
-using Azure.Provisioning;
-using Azure.Provisioning.Expressions;
-using Azure.Provisioning.Primitives;
-using Azure.Provisioning.Roles;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
 
 namespace Azure.Projects;
