+++ sdk/cloudmachine/Azure.Projects.Provisioning/src/Tooling/AzureProjectsCommands.cs
@@ -3,9 +3,6 @@
 
 using System;
 using System.ClientModel.Primitives;
-using System.ClientModel.TypeSpec;
-using System.Collections.Generic;
-using System.IO;
 using System.Text.Json;
 using Azure.Core;
 using Azure.Core.Rest;
@@ -40,12 +37,6 @@ public static class AzureProjectsCommands
             Azd.InitDeployment(cmi, projName);
             return true;
         }
-
-        //if (args[0] == "-tsp")
-        //{
-        //    GenerateTsp(cmi.Endpoints);
-        //    return true;
-        //}
 
         if (args[0] == "-ai")
         {
@@ -116,23 +107,6 @@ public static class AzureProjectsCommands
                         Console.WriteLine($"{deprecatedAt} | {modelId}");
                 }
             }
-        }
-    }
-
-    private static void GenerateTsp(IEnumerable<Type> operationGroups)
-    {
-        foreach (Type operationGroup in operationGroups)
-        {
-            string name = operationGroup.Name;
-            if (name.StartsWith("I"))
-                name = name.Substring(1);
-            string directory = Path.Combine(".", "tsp");
-            string tspFile = Path.Combine(directory, $"{name}.tsp");
-            Directory.CreateDirectory(Path.GetDirectoryName(tspFile)!);
-            if (File.Exists(tspFile))
-                File.Delete(tspFile);
-            using FileStream stream = File.OpenWrite(tspFile);
-            TypeSpecWriter.WriteServer(stream, operationGroup);
         }
     }
 }
