+++ sdk/cloudmachine/Azure.Projects.Web/src/Tsp/TspEndpoints.cs
@@ -6,9 +6,18 @@ using System.Collections.Generic;
 
 namespace Azure.Projects;
 
+/// <summary>
+/// This class is used to register TSP endpoints.
+/// </summary>
 public class TspEndpoints
 {
     internal List<Type> Endpoints { get; } = [];
+
+    /// <summary>
+    /// Adds a TSP endpoint to the list of endpoints.
+    /// </summary>
+    /// <typeparam name="T"></typeparam>
+    /// <exception cref="InvalidOperationException"></exception>
     public void AddEndpoints<T>()
     {
         Type endpointsType = typeof(T);
