+++ sdk/cloudmachine/Azure.Projects.sln
@@ -31,6 +31,13 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HelloRAG", "Azure.Projects\
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OFXRAG", "Azure.Projects\samples\OFXRAG\OFXRAG.csproj", "{56CDEC41-54BA-89E5-DF48-4A2C60F4C2F4}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Projects.Web.Tests", "Azure.Projects.Web\tests\Azure.Projects.Web.Tests.csproj", "{CD951F82-21B7-C6FC-B80E-A2EBBB2A75F9}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{8EC462FD-D22E-90A8-E5CE-7E832BA40C5D}"
+	ProjectSection(SolutionItems) = preProject
+		README.md = README.md
+	EndProjectSection
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -89,6 +96,10 @@ Global
 		{56CDEC41-54BA-89E5-DF48-4A2C60F4C2F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{56CDEC41-54BA-89E5-DF48-4A2C60F4C2F4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{56CDEC41-54BA-89E5-DF48-4A2C60F4C2F4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CD951F82-21B7-C6FC-B80E-A2EBBB2A75F9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CD951F82-21B7-C6FC-B80E-A2EBBB2A75F9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CD951F82-21B7-C6FC-B80E-A2EBBB2A75F9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CD951F82-21B7-C6FC-B80E-A2EBBB2A75F9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
