+++ sdk/cloudmachine/Azure.Projects/samples/HelloWorld/Program.cs
@@ -8,7 +8,7 @@ using OpenAI.Chat;
 ProjectInfrastructure infrastructure = new();
 infrastructure.AddFeature(new OpenAIModelFeature("gpt-35-turbo", "0125"));
 
-// the app can be called with -init switch to generate bicep and prepare for azd deployment.
+// the app can be called with -bicep or -init switch to generate bicep and prepare for azd deployment.
 if (infrastructure.TryExecuteCommand(args)) return;
 
 ProjectClient project = infrastructure.GetClient();
