+++ sdk/cloudmachine/Azure.Projects/src/AIFoundary/AIFoundryClient.cs
@@ -5,17 +5,10 @@ using System;
 using System.ClientModel.Primitives;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Linq;
 using System.Threading;
-using Azure.AI.Inference;
-using Azure.AI.OpenAI;
 using Azure.AI.Projects;
 using Azure.Core;
 using Azure.Identity;
-using Azure.Search.Documents;
-using Azure.Search.Documents.Indexes;
-using OpenAI.Chat;
-using OpenAI.Embeddings;
 
 namespace Azure.Projects
 {
