+++ sdk/cloudmachine/Azure.Projects/src/Core/RestCLient/RestClient.cs
@@ -2,8 +2,8 @@
 // Licensed under the MIT License.
 
 using System;
-using System.ClientModel.Primitives;
 using System.ClientModel;
+using System.ClientModel.Primitives;
 
 namespace Azure.Core.Rest
 {
